Moves Event related component to dedicated folder
